moved CountrySelect component under PageLayout component
