+++ src/client/components/PageLayout/Footer/Footer.tsx
@@ -1,4 +1,4 @@
-import './Footer.scss'
+import './footer.scss'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
